+++ internal/dial/conn.go
@@ -12,10 +12,11 @@ type netConn struct {
 	address string
 	trace   trace.Driver
 	cc      net.Conn
+	ctx     context.Context
 }
 
 func newConn(ctx context.Context, address string, t trace.Driver) (_ net.Conn, err error) {
-	onDone := trace.DriverOnNetDial(t, address)
+	onDone := trace.DriverOnNetDial(t, ctx, address)
 	defer func() {
 		onDone(err)
 	}()
@@ -27,11 +28,12 @@ func newConn(ctx context.Context, address string, t trace.Driver) (_ net.Conn, e
 		address: address,
 		cc:      cc,
 		trace:   t,
+		ctx:     ctx,
 	}, nil
 }
 
 func (c *netConn) Read(b []byte) (n int, err error) {
-	onDone := trace.DriverOnNetRead(c.trace, c.address, len(b))
+	onDone := trace.DriverOnNetRead(c.trace, c.ctx, c.address, len(b))
 	defer func() {
 		onDone(n, err)
 	}()
@@ -39,7 +41,7 @@ func (c *netConn) Read(b []byte) (n int, err error) {
 }
 
 func (c *netConn) Write(b []byte) (n int, err error) {
-	onDone := trace.DriverOnNetWrite(c.trace, c.address, len(b))
+	onDone := trace.DriverOnNetWrite(c.trace, c.ctx, c.address, len(b))
 	defer func() {
 		onDone(n, err)
 	}()
@@ -47,7 +49,7 @@ func (c *netConn) Write(b []byte) (n int, err error) {
 }
 
 func (c *netConn) Close() (err error) {
-	onDone := trace.DriverOnNetClose(c.trace, c.address)
+	onDone := trace.DriverOnNetClose(c.trace, c.ctx, c.address)
 	defer func() {
 		onDone(err)
 	}()